Process enviiornment
Working on a process environment to make them stand out typographically.  Overall this environment looks nice, though I need to put figures in a white box so that the background color does not leak through the includegraphics object.  Todo: need to typeset the captions for figures left justified. Having a huge issue trying to make a TOC for the processes.